pyproject.toml
[tool.poetry]
name = "appman"
version = "0.1.0"
description = "Application Management"
authors = ["Basilio Bogado <541149+basiliskus@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
click = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
yamale = "^3.0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
